requirements.txt
Django==5.2.8
pycryptodome>=3.20
Pillow>=10.0